projects/ai-meeting-notes-generator/requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0

# AI/ML
openai-whisper>=20231117
torch>=2.0.0

# HTTP & API
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# CORS
starlette>=0.27.0
